Prepare for 2.6.1 release

Approved-by: Louise Poubel <lupoubel@hotmail.com>

+++ src/systems/battery_plugin/LinearBatteryPlugin.cc
@@ -295,7 +295,7 @@ void LinearBatteryPlugin::PreUpdate(
       if (jointVelocityCmd) {
         for (double jointVel : jointVelocityCmd->Data())
         {
-          if (fabsf(jointVel) > 0)
+          if (fabsf(static_cast<float>(jointVel)) > 0)
           {
             this->dataPtr->startDraining = true;
             return;
